The first "release" of the app.
The app has been somewhat completed and is now usable. Due to an issue where most apps would block automated input, I've decided to get Java to help me automate input. However, this resulted in fast key presses being too slow and not being able to catch up with the player as each key press is a separate Java process that needs to be dealt with. To work around this, I am going to have to code the player inside of the Java helper jar file as well. The app is far from finished when in comes to back-end stuff. It is not recommended to use this "release", meanwhile I will commit this just for cloud back-up reasons.